+++ inc/ti/fn/fnimport.h
@@ -98,29 +98,31 @@ static int do__f_import(ti_query_t * query, cleri_node_t * nd, ex_t * e)
     if (ti_task_add_import(task, bytes, import_tasks))
     {
         ex_set_mem(e);  /* might overwrite EX_BAD_DATA */
+        goto fail0;
     }
-    else if (!import_tasks)
+
+    /* we now have a new collection root, thus a new task list; this new list
+     * is required for the optional tasks but also in case other statements
+     * follow the import function */
+    task = ti_task_new_task(query->change, query->collection->root);
+    if (!task)
+    {
+        ex_set_mem(e);
+        goto fail0;
+    }
+
+    if (!import_tasks)
     {
         ti_collection_tasks_clear(query->collection);
     }
-    else
+    else for (vec_each(query->collection->vtasks, ti_vtask_t, vtask))
     {
-        /* this is a new collection root, thus a new task list */
-        task = ti_task_new_task(query->change, query->collection->root);
-        if (!task)
-        {
-            ex_set_mem(e);
-            goto fail0;
-        }
-        for (vec_each(query->collection->vtasks, ti_vtask_t, vtask))
-        {
-            /* get ownership of all the tasks */
-            ti_user_drop(vtask->user);
-            vtask->user = query->user;
-            ti_incref(query->user);
-            if (ti_task_add_vtask_set_owner(task, vtask))
-                ex_set_mem(e);  /* task cleanup is not required */
-        }
+        /* get ownership of all the tasks */
+        ti_user_drop(vtask->user);
+        vtask->user = query->user;
+        ti_incref(query->user);
+        if (ti_task_add_vtask_set_owner(task, vtask))
+            ex_set_mem(e);  /* task cleanup is not required */
     }
 
 fail0: